+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/models/VariableSet.java
@@ -25,7 +25,7 @@ import java.util.Set;
 public class VariableSet {
 
     //TODO Is this field really needed?
-    private int id;
+    private long id;
     private String name;
     private boolean unique;
     private String description;
@@ -36,7 +36,7 @@ public class VariableSet {
     public VariableSet() {
     }
 
-    public VariableSet(int id, String name, boolean unique, String description, Set<Variable> variables,
+    public VariableSet(long id, String name, boolean unique, String description, Set<Variable> variables,
                        Map<String, Object> attributes) {
         this.id = id;
         this.name = name;
@@ -59,11 +59,11 @@ public class VariableSet {
         return sb.toString();
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
